+++ src/seamlesserverpkg/renderer/routes/default/notFound.nim
@@ -1,9 +1,8 @@
 include pkg/karax/prelude
 
-import ../../state
 import ../../base
 
-proc renderHtml*(state = newState()): Rendered =
+proc renderHtml*(state: State): Rendered =
   ## Renders the HTML for homepage
   new result
   result.title = "Not Found"
@@ -15,4 +14,4 @@ when not defined js:
 
   proc render*(ctx: Context) {.async.} =
     ## Server side homepage renderer
-    resp renderHtml().inBaseHtml
+    resp ssr renderHtml