+++ src/seamlesserverpkg/renderer/routes/home.nim
@@ -1,9 +1,10 @@
 include pkg/karax/prelude
 
-import ../state
 import ../base
 
-proc renderHtml*(state = newState()): Rendered =
+const path* = "/"
+
+proc renderHtml*(state: State): Rendered =
   ## Renders the HTML for homepage
   new result
   result.title = "Home"
@@ -11,10 +12,13 @@ proc renderHtml*(state = newState()): Rendered =
     h1: text "homepage"
     dynamicLink(href = "/user/login"):
       text "Login"
+    br()
+    dynamicLink(href = "/user/register"):
+      text "Register"
 
 when not defined js:
   import pkg/prologue
 
   proc render*(ctx: Context) {.async.} =
     ## Server side homepage renderer
-    resp renderHtml().inBaseHtml
+    resp ssr renderHtml