+++ src/seamlesserverpkg/renderer/state.nim
@@ -1,7 +1,0 @@
-type
-  State* = ref object
-    ## State of frontend page
-
-func newState*: State =
-  ## Creates new state
-  new result